Move the constructor of Shape to the factory
