--- src/Hilke.KineticConvolution/DirectionHelper.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace Hilke.KineticConvolution
-{
-    internal static class DirectionHelper
-    {
-    }
-}
