+++ src/Hilke.KineticConvolution/Point.cs
@@ -7,7 +7,7 @@ namespace Hilke.KineticConvolution
     {
         private readonly IAlgebraicNumberCalculator<TAlgebraicNumber> _calculator;
 
-        public Point(IAlgebraicNumberCalculator<TAlgebraicNumber> calculator, TAlgebraicNumber x, TAlgebraicNumber y)
+        internal Point(IAlgebraicNumberCalculator<TAlgebraicNumber> calculator, TAlgebraicNumber x, TAlgebraicNumber y)
         {
             _calculator = calculator ?? throw new ArgumentNullException(nameof(calculator));
             X = x ?? throw new ArgumentNullException(nameof(x));
